post-setting - Post 모델 수정, serializer 디렉토리 구조수정(import 오류때문에 수정)
