+++ django_app/diary/models.py
@@ -12,7 +12,7 @@ class Diary(models.Model):
 
 
 class Post(models.Model):
-    diary = models.ForeignKey(Diary)
+    diary = models.ForeignKey(Diary, on_delete=models.CASCADE)
     content = models.TextField()
     created_date = models.DateTimeField(auto_now_add=True)
 
@@ -21,7 +21,7 @@ class Post(models.Model):
 
 
 class PostPhoto(models.Model):
-    post = models.ForeignKey(Post)
+    post = models.ForeignKey(Post, on_delete=models.CASCADE)
     photo = models.ImageField(upload_to='post')
     gpsLatitude = models.FloatField("Latitude", blank=True, null=True)
     gpsLongitude = models.FloatField("Longitude", blank=True, null=True)
