+++ django_app/diary/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .diary import *
+from .post import *
+from .post_photo import *
